refactor to delete layout file and move setup to App.tsx
