--- frontend/.react-router/types/app/layouts/+types/default.ts
@@ -1,62 +1,0 @@
-// Generated by React Router
-
-import type { GetInfo, GetAnnotations } from "react-router/internal";
-
-type Module = typeof import("../default.js")
-
-type Info = GetInfo<{
-  file: "layouts/default.tsx",
-  module: Module
-}>
-
-type Matches = [{
-  id: "root";
-  module: typeof import("../../root.js");
-}, {
-  id: "layouts/default";
-  module: typeof import("../default.js");
-}];
-
-type Annotations = GetAnnotations<Info & { module: Module, matches: Matches }, false>;
-
-export namespace Route {
-  // links
-  export type LinkDescriptors = Annotations["LinkDescriptors"];
-  export type LinksFunction = Annotations["LinksFunction"];
-
-  // meta
-  export type MetaArgs = Annotations["MetaArgs"];
-  export type MetaDescriptors = Annotations["MetaDescriptors"];
-  export type MetaFunction = Annotations["MetaFunction"];
-
-  // headers
-  export type HeadersArgs = Annotations["HeadersArgs"];
-  export type HeadersFunction = Annotations["HeadersFunction"];
-
-  // middleware
-  export type MiddlewareFunction = Annotations["MiddlewareFunction"];
-
-  // clientMiddleware
-  export type ClientMiddlewareFunction = Annotations["ClientMiddlewareFunction"];
-
-  // loader
-  export type LoaderArgs = Annotations["LoaderArgs"];
-
-  // clientLoader
-  export type ClientLoaderArgs = Annotations["ClientLoaderArgs"];
-
-  // action
-  export type ActionArgs = Annotations["ActionArgs"];
-
-  // clientAction
-  export type ClientActionArgs = Annotations["ClientActionArgs"];
-
-  // HydrateFallback
-  export type HydrateFallbackProps = Annotations["HydrateFallbackProps"];
-
-  // Component
-  export type ComponentProps = Annotations["ComponentProps"];
-
-  // ErrorBoundary
-  export type ErrorBoundaryProps = Annotations["ErrorBoundaryProps"];
-}
