+++ frontend/app/routes.ts
@@ -1,8 +1,6 @@
-import { type RouteConfig, index, layout, route } from "@react-router/dev/routes";
+import { type RouteConfig, index, route } from "@react-router/dev/routes";
 
 export default [
-    layout("layouts/default.tsx", [
-        index("routes/home.tsx"),
-        route("forecast/:locationId", "routes/forecast.tsx")
-    ])
+    index("routes/home.tsx"),
+    route("forecast/:locationId", "routes/forecast.tsx")
 ] satisfies RouteConfig;
